pyproject.toml
[tool.poetry]
name = "telegram-stats-bot"
version = "0.8.1"
description = "A logging and statistics bot for Telegram based on python-telegram-bot."
authors = ["Michael DM Dryden <mk.dryden@utoronto.ca>"]
repository = "https://github.com/mkdryden/telegram-stats-bot"
license = "gpl-3.0-or-later"
readme = "README.rst"
include = ["examples/*", "CHANGELOG.rst"]

[tool.poetry.dependencies]
python = ">=3.9, <3.10"
python-telegram-bot = {version = "^20.6", extras = ["job-queue"]}
SQLAlchemy = "^2.0.23"
SQLAlchemy-Utils = "^0.41"
pandas = "^2.1"
pandas-stubs = "^2.1"
seaborn = "^0.13"
numpy = "^1.26.1"
matplotlib = "^3.8"
appdirs = "^1.4.4"
single-source = "^0.3.0"
typer = "^0.9.0"
psycopg = {version = "^3.1.12", extras = ["binary"]}
pytest = "^7.4.3"
alembic = "^1.13.3"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
pytest-postgresql = "^5.0.0"
random-word = "^1.0.11"

[tool.pyright]
venv=".venv"
venvPath="."
pythonVersion="3.9.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
